pyproject.toml
[tool.poetry]
name = "gh-echo"
version = "0.3.0"
description = "A command-line tool built to obtain in-depth, actionable information about GitHub repositories."
authors = ["Aryan Khurana <aryankhurana1511@gmail.com>"]
license = "MIT"
readme = "README.md"
homepage = "https://pypi.org/project/gh-echo/o"
repository = "https://github.com/AryanK1511/github-echo"
documentation = "https://github.com/AryanK1511/github-echo/blob/main/README.md"
keywords = ["GitHub", "GenAI", "Python3", "CLI"]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent"
]
packages = [{include = "application"}]
include = ["_main.py", "_config.py", "pyproject.toml"]

[tool.poetry.scripts]
lint = "_scripts:lint" # Lint all files in the current directory, and fix any fixable errors.
format = "_scripts:format_code" # Format all files in the current directory
lint-and-format = "_scripts:lint_and_format" # Lint and format
gh-echo = "_main:app"

[tool.poetry.dependencies]
python = "^3.9"
typer = "^0.12.5"
google-generativeai = "^0.8.0"
python-dotenv = "^1.0.1"
httpx = "^0.27.2"
pytest = "^8.3.3"
single-source = "^0.4.0"
groq = "^0.11.0"
toml = "^0.10.2"
pre-commit = "^4.0.1"

[tool.poetry.group.dev.dependencies]
ruff = "^0.6.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"


